Задание 4

+++ src/TaskWork/TaskWork1/TaskWork4.java
@@ -10,17 +10,15 @@ public class TaskWork4 {
 
         int a = one.nextInt();
 
-
         System.out.println("Введите второе число");
 
         int b = one.nextInt();
-
 
         System.out.println("Введите третье число");
 
         int c = one.nextInt();
 
-        int d = (a + b + c )/ 3;
+        double d = (a + b + c )/ 3;
         System.out.println("Среднее арифметическое трех числе " + d);
 
         if(d/2 > 3){
